Moved script files to Runtime directory (left metas untouched for now)
