+++ src/Ros2ForUnity/Runtime/Scripts/ROS2ListenerExample.cs
@@ -1,0 +1,46 @@
+// Copyright 2019-2021 Robotec.ai.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using System;
+using UnityEngine;
+
+namespace ROS2
+{
+
+/// <summary>
+/// An example class provided for testing of basic ROS2 communication
+/// </summary>
+public class ROS2ListenerExample : MonoBehaviour
+{
+    private ROS2UnityComponent ros2Unity;
+    private ROS2Node ros2Node;
+    private ISubscription<std_msgs.msg.String> chatter_sub;
+
+    void Start()
+    {
+        ros2Unity = GetComponent<ROS2UnityComponent>();
+    }
+
+    void Update()
+    {
+        if (ros2Node == null && ros2Unity.Ok())
+        {
+            ros2Node = ros2Unity.CreateNode("ROS2UnityListenerNode");
+            chatter_sub = ros2Node.CreateSubscription<std_msgs.msg.String>(
+              "chatter", msg => Debug.Log("Unity listener heard: [" + msg.Data + "]"));
+        }
+    }
+}
+
+}  // namespace ROS2
